+++ webfx-kit/webfx-kit-javafxgraphics-emul/src/main/java/javafx/animation/ParallelTransition.java
@@ -545,7 +545,7 @@ public final class ParallelTransition extends Transition {
 
                     offsetTicks[i] = (getCurrentRate() > 0)? newTicks - delays[i] : add(durations[i], delays[i]) - newTicks;
                 } else if (status == Status.PAUSED) {
-                    offsetTicks[i] += (newTicks - oldTicks) * Math.signum(this.clipEnvelope.getCurrentRate());
+                    offsetTicks[i] += (/*cast to fix CodeQL implicit-cast-in-compound-assignment security issue*/long) ((newTicks - oldTicks) * Math.signum(this.clipEnvelope.getCurrentRate()));
                 } else {
                     offsetTicks[i] += (getCurrentRate() > 0) ? newTicks - oldTicks : oldTicks - newTicks;
                 }
